+++ packages/ui/src/index.ts
@@ -4,5 +4,6 @@ export * from "./branding";
 export * from "./components";
 export * from "./icons/Icon";
 export * from "./icons/names";
+export * from "./modules";
 
 export const UI_PACKAGE_NAME = "@agilityhub/ui" as const;